+++ docs/change_log.rst
@@ -72,11 +72,12 @@ The following are only available within Stem's `git repository
  * **Utilities**
 
   * IPv6 support in :func:`~stem.util.connection.get_connections` when resolving with proc, netstat, lsof, or ss (:trac:`18079`)
-  * Added :func:`~stem.util.__init__.datetime_to_unix`
   * The 'ss' connection resolver didn't work on Gentoo (:trac:`18079`)
   * Recognize IPv4-mapped IPv6 addresses in our utils (:trac:`18079`)
-  * Added an **is_ipv6** value to :class:`~stem.util.connection.Connection` instances
   * Allow :func:`stem.util.conf.Config.set` to remove values when provided with a **None** value
+  * Additional information when :func:`~stem.util.system.call` fails through a :class:`~stem.util.system.CallError`
+  * Added an **is_ipv6** value to :class:`~stem.util.connection.Connection` instances
+  * Added :func:`~stem.util.__init__.datetime_to_unix`
 
  * **Interpreter**